+++ CHANGELOG.md
@@ -14,6 +14,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - database/sql: DB connection configuration
 - database/sql: Utility function to make the master connection the primary one
 
+### Fixed
+
+- database/sql: Slave overwriting master connection
+
 
 ## [0.17.0] - 2017-02-04