+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+
+## [0.18.0] - 2018-02-28
+
 ### Added
 
 - database/sql: DB connection configuration
@@ -19,7 +22,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - database/sql: Slave overwriting master connection
 
 
-## [0.17.0] - 2017-02-04
+## [0.17.0] - 2018-02-04
 
 ### Added
 
@@ -294,7 +297,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Bootstrap options
 
 
-[Unreleased]: https://github.com/goph/fxt/compare/v0.17.0...HEAD
+[Unreleased]: https://github.com/goph/fxt/compare/v0.18.0...HEAD
+[0.18.0]: https://github.com/goph/fxt/compare/v0.17.0...v0.18.0
 [0.17.0]: https://github.com/goph/fxt/compare/v0.16.0...v0.17.0
 [0.16.0]: https://github.com/goph/fxt/compare/v0.15.1...v0.16.0
 [0.15.1]: https://github.com/goph/fxt/compare/v0.15.0...v0.15.1